feat(components): add coming-soon page component
- Create standalone ComingSoonPageComponent
- Add basic template and styling for coming-soon page
- Configure RouterOutlet in AppComponent
- Update app configuration for proper route handling

BREAKING CHANGE: AppComponent now requires RouterOutlet import
