+++ src/app/temp/comming-soon-page/comming-soon-page.component.html
@@ -1,0 +1,38 @@
+<div class="banner-container">
+    <div class="banner-content">
+      <h1>Prime Market <span>Terminal</span></h1>
+      <p class="platform-label">Institutional-Grade Trading Platform</p>
+      
+      <!-- Shields.io Badges -->
+      <div class="badge-container">
+        <img src="https://img.shields.io/badge/Angular-19.0.0-dd0031?logo=angular" alt="Angular" />
+        <img src="https://img.shields.io/badge/TypeScript-5.4-blue?logo=typescript" alt="TypeScript" />
+        <img src="https://img.shields.io/badge/Status-BETA-orange" alt="Status" />
+        <img src="https://img.shields.io/badge/build-passing-brightgreen.svg" alt="Build Status" />
+        <img src="https://img.shields.io/badge/PRs-welcome-brightgreen.svg" alt="PRs Welcome" />
+      </div>
+  
+      <!-- Technology Stack -->
+      <div class="tech-stack">
+        @for (icon of techIcons; track icon.src) {
+          <div class="tech-icon-wrapper">
+            <img [src]="icon.src" [alt]="icon.alt" [title]="icon.alt" />
+          </div>
+        }
+      </div>
+  
+      <p class="subtitle">
+        Empowering traders with real-time market data and advanced analytics using
+        <code>Angular</code> and cutting-edge technologies
+      </p>
+  
+      <!-- Development Notice -->
+      <div class="development-notice">
+        <p class="coming-soon">Coming Soon</p>
+        <p class="dev-message">
+          We're working hard to bring you the next generation of trading technology.
+          Stay tuned for updates!
+        </p>
+      </div>
+    </div>
+  </div>
