+++ src/app/temp/comming-soon-page/comming-soon-page.component.scss
@@ -1,0 +1,183 @@
+// Import variables
+@use '../../../scss/base/variables' as *;
+// Import mixins
+@use '../../../scss/base/mixins' as *;
+
+// Map of technology icons used in the application, mapping technology names to their CDN URLs
+$tech-icons: (
+  'angular': 'https://cdn.jsdelivr.net/gh/devicons/devicon/icons/angular/angular-original.svg',
+  'typescript': 'https://cdn.jsdelivr.net/gh/devicons/devicon/icons/typescript/typescript-original.svg',
+  'sass': 'https://cdn.jsdelivr.net/gh/devicons/devicon/icons/sass/sass-original.svg',
+  'nodejs': 'https://cdn.jsdelivr.net/gh/devicons/devicon/icons/nodejs/nodejs-original.svg',
+  'github': 'https://cdn.jsdelivr.net/gh/devicons/devicon/icons/github/github-original-wordmark.svg'
+);
+
+.banner-container {
+  min-height: 100vh;
+  width: 100vw;
+  padding: 2rem;
+  background: linear-gradient(135deg, #1a1a1a 0%, #2d2d2d 100%);
+  color: $white;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  overflow: hidden;
+
+  .banner-content {
+    max-width: 1200px;
+    margin: 0 auto;
+    text-align: center;
+  }
+
+  h1 {
+    font-size: 4rem;
+    margin-bottom: 0.5rem;
+    font-weight: 700;
+
+    span {
+      color: $primary;
+      position: relative;
+
+      &::after {
+        content: '';
+        position: absolute;
+        bottom: -5px;
+        left: 0;
+        width: 100%;
+        height: 2px;
+        background: $primary;
+        transform: scaleX(0);
+        animation: underline 1.5s ease-in-out forwards;
+      }
+    }
+  }
+
+  .platform-label {
+    font-size: 1.5rem;
+    margin-bottom: 1.5rem;
+    color: rgba($white, 0.9);
+  }
+
+  .badge-container {
+    display: flex;
+    gap: 0.5rem;
+    justify-content: center;
+    margin: 1.5rem 0;
+    flex-wrap: wrap;
+
+    img {
+      height: 20px;
+      transition: transform 0.2s ease;
+
+      &:hover {
+        transform: translateY(-2px);
+      }
+    }
+  }
+
+  .subtitle {
+    font-size: 1.2rem;
+    margin: 2rem 0;
+    color: rgba($white, 0.8);
+
+    code {
+      background: rgba($primary, 0.1);
+      color: $primary;
+      padding: 0.2rem 0.5rem;
+      border-radius: 4px;
+    }
+  }
+
+  .development-notice {
+    @include backdrop-blur(10px);
+    margin-top: 3rem;
+    padding: 1rem 2rem;
+    background: rgba($white, 0.05);
+    border-radius: 8px;
+
+    .coming-soon {
+      font-size: 0.9rem;
+      text-transform: uppercase;
+      letter-spacing: 2px;
+      color: $primary;
+      margin-bottom: 0.5rem;
+    }
+
+    .dev-message {
+      font-size: 1rem;
+      color: rgba($white, 0.7);
+    }
+  }
+
+  .tech-stack {
+    display: flex;
+    gap: 2rem;
+    align-items: center;
+    justify-content: center;
+    padding: 2rem 0;
+
+    .tech-icon-wrapper {
+      position: relative;
+      width: 48px;
+      height: 48px;
+      transition: transform 0.3s ease;
+
+      &:hover {
+        transform: translateY(-3px);
+
+        &::after {
+          opacity: 0.4;
+          transform: scale(1.1);
+        }
+      }
+
+      &::after {
+        content: '';
+        position: absolute;
+        top: 0;
+        left: 0;
+        width: 100%;
+        height: 100%;
+        background-size: contain;
+        background-position: center;
+        background-repeat: no-repeat;
+        filter: blur(12px) brightness(1.2);
+        opacity: 0;
+        transition: all 0.3s ease;
+        z-index: -1;
+      }
+
+      $index: 1;
+      @each $name, $url in $tech-icons {
+        &:nth-child(#{$index})::after {
+          background-image: url('#{$url}');
+        }
+        $index: $index + 1;
+      }
+
+      img {
+        width: 100%;
+        height: 100%;
+        object-fit: contain;
+        position: relative;
+        z-index: 1;
+      }
+    }
+  }
+}
+
+@keyframes underline {
+  to {
+    transform: scaleX(1);
+  }
+}
+
+@media (max-width: 768px) {
+  .banner-container {
+    padding: 1rem;
+
+    h1 {
+      font-size: 2.5rem;
+    }
+  }
+}
